requirements.txt
flask-marshmallow
psycopg2-binary
transformers
flask-sqlalchemy
marshmallow-sqlalchemy
flask
flask_restful
torch
peft
flask-bcrypt
flask-wtf
email_validator
flask-login
flask-mail
flask-mailing
